+++ QuantumRelay/QuantumRelayInfo.cs
@@ -3,7 +3,7 @@ namespace QuantumRelay
     internal static class QuantumRelayInfo
     {
         internal const string ModName = "Quantum Relay";
-        internal const string Version = "1.6.0-preview1";
+        internal const string Version = "1.6.0-preview2";
         internal const string Developer = "SockedRooster";
         internal const string Company = "RoosterWorks";
         internal const string License = "MIT";